+++ api/arch.hpp
@@ -29,6 +29,11 @@ extern void __arch_poweroff();
 extern void __arch_reboot();
 extern void __arch_enable_legacy_irq(uint8_t);
 extern void __arch_disable_legacy_irq(uint8_t);
+
+extern void __arch_install_irq(uint8_t, void(*)());
+extern void __arch_subscribe_irq(uint8_t);
+extern void __arch_unsubscribe_irq(uint8_t);
+
 inline void __arch_hw_barrier() noexcept;
 inline void __sw_barrier() noexcept;
 inline uint64_t __arch_cpu_cycles() noexcept;
