+++ src/platform/x86_nano/CMakeLists.txt
@@ -12,6 +12,7 @@ set(PLATFORM_OBJECTS
   ../x86_pc/block.cpp
   ../x86_pc/softreset.cpp
   ../x86_pc/sanity_checks.cpp
+  ../x86_pc/idt.cpp
   platform.cpp
   kernel_start.cpp
   )
